refactor: unify Coupang Excel parsing
