+++ routes/api/coupangApi.js
@@ -1,8 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const multer = require("multer");
-const xlsx = require("xlsx");
-const safeReadXlsx = require("../../lib/safeReadXlsx");
+const parseCoupangExcel = require("../../lib/parseCoupangExcel");
 const fs = require("fs");
 const path = require("path");
 
@@ -10,38 +9,13 @@ const uploadsDir = path.join(__dirname, "../../uploads");
 if (!fs.existsSync(uploadsDir)) fs.mkdirSync(uploadsDir);
 const upload = multer({ dest: uploadsDir });
 
-function parseExcel(filePath) {
-  const workbook = safeReadXlsx(filePath);
-  const sheet = workbook.Sheets[workbook.SheetNames[0]];
-  const rows = xlsx.utils.sheet_to_json(sheet, { header: 1 }).slice(2);
-  return rows
-    .map((row) => {
-      const obj = {};
-      obj["Option ID"] = String(row[2] ?? "").trim();
-      obj["Product name"] = row[4] ?? "";
-      obj["Option name"] = row[5] ?? "";
-      const inventory = Number(String(row[7]).replace(/,/g, "")) || 0;
-      obj["Orderable quantity (real-time)"] = inventory;
-      const salesAmount = Number(String(row[11]).replace(/,/g, "")) || 0;
-      obj["Sales amount on the last 30 days"] = salesAmount;
-      const salesCount = Number(String(row[13]).replace(/,/g, "")) || 0;
-      obj["Sales in the last 30 days"] = salesCount;
-      const daily = salesCount / 30;
-      const safety = daily * 7;
-      obj["Shortage quantity"] =
-        inventory < safety ? Math.ceil(safety - inventory) : 0;
-      return obj;
-    })
-    .filter((item) => item["Option ID"]);
-}
-
 router.post("/upload", upload.single("excelFile"), async (req, res) => {
   const db = req.app.locals.db;
   try {
     if (!req.file)
       return res.status(400).json({ status: "error", message: "파일이 없습니다." });
     const filePath = req.file.path;
-    const data = parseExcel(filePath);
+    const data = parseCoupangExcel(filePath);
     const bulkOps = data.map((item) => ({
       updateOne: {
         filter: { "Option ID": item["Option ID"] },
